feat: splitAndSum 메서드의 입력으로 "//"와 "\n" 문자 사이에 커스텀 구분자를 지정한 경우, 해당 커스텀 구분자를 기준으로 구분하고 숫자들의 합을 반환
