+++ README.md
@@ -16,6 +16,6 @@
 - [X] ~~빈 문자열 또는 null 값인 경우 0을 반환 ("" => 0, null => 0)~~
 - [X] ~~숫자 두개를 컴마(,) 구분자로 입력할 경우 두 숫자의 합을 반환("1,2"=>3)~~
 - [X] ~~구분자를 컴마(,) 이외에 콜론(:)을 사용할 수 있다.("1,2:3" => 6)~~
-- [ ] "//"와 "\n" 문자 사이에 커스텀 구분자를 지정할 수 있다.("//;\n1;2;3" => 6)
+- [X] ~~"//"와 "\n" 문자 사이에 커스텀 구분자를 지정할 수 있다.("//;\n1;2;3" => 6)~~
 - [X] ~~음수를 전달할 경우 RuntimeException 예외 발생("-1,2,3" => RuntimeException)~~
 - [X] ~~숫자 하나를 문자열로 입력할 경우 해당 숫자를 반환 ("1" => 1)~~
