+++ src/main/java/calculator/StringAddCalculator.java
@@ -1,16 +1,10 @@
 package calculator;
 
+import parser.Parser;
+
 public class StringAddCalculator {
     public static int splitAndSum(final String input) {
-        if (isEmpty(input)) {
-            return 0;
-        }
-
-        return sum(toInts(split(input)));
-    }
-
-    private static String[] split(final String input) {
-        return input.split("[,|:]");
+        return sum(Parser.parse(input));
     }
 
     private static int sum(final int[] numbers) {
@@ -21,25 +15,5 @@ public class StringAddCalculator {
         }
 
         return sum;
-    }
-
-    private static int[] toInts(final String[] numbers) {
-        int[] intNumbers = new int[numbers.length];
-
-        for (int i = 0; i < numbers.length; i++) {
-            final int parseInt = Integer.parseInt(numbers[i]);
-
-            if (parseInt < 0) {
-                throw new RuntimeException("Input must be a positive integer");
-            }
-
-            intNumbers[i] = parseInt;
-        }
-
-        return intNumbers;
-    }
-
-    private static boolean isEmpty(final String input) {
-        return input == null || input.isEmpty();
     }
 }
